training/requirements.txt
numpy
setuptools
tqdm
xmltodict
jsonlines
torch
torchvision
lightning
wandb
transformers
datasets
evaluate
pillow
Pygments
